+++ src/vscode/treeviews/Gateways/iPowerBIGatewayItem.ts
@@ -1,14 +1,11 @@
 import { UniqueId } from '../../../helpers/Helper';
+import { PowerBIApiTreeItem } from '../PowerBIApiTreeItem';
 import { ApiItemType } from '../_types';
 
-export interface iPowerBIGatewayItem {
+export interface iPowerBIGatewayItem extends PowerBIApiTreeItem {
 	/* from iPowerBIApiItem */
 	itemType: ApiItemType;
 	uid: UniqueId;
 	id?: string;
 	name: string;
-
-	/* new properties */
-	type: string;
-	publicKey: object;
 }