+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove Fabric internal dataflows without a name ([#13](/../../issues/13))
 - Added `Copy ... to Clibpoard` commands to the context menu ([#14](/../../issues/14))
 - Added `Open in Power BI Service` command to the context menu
+- Added Gateway Data Sources to Gateway treeview
 
 **v1.0.0**:
 - added support for TMDL (see [TMDL requirements](/README.md#tmdl-requirements))